requirements.txt
# The Sentinel: Multimodal Alpha Engine
# Core dependencies

# Database
duckdb>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# News Scraping
feedparser>=6.0.0
requests>=2.31.0

# Gemini AI
google-generativeai>=0.5.0

# Charting
matplotlib>=3.7.0
mplfinance>=0.12.9b7

# Scheduling & Utils
schedule>=1.2.0
python-dotenv>=1.0.0

# Async Support
aiohttp>=3.9.0

# Web Dashboard
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
